(#7) Clean up and add ListItem for next story
Added clauses for PropertyController.

Deleted ExampleComponent and added ListItem for next story.

+++ app/Http/Controllers/Api/PropertyController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Property;
 use Auth;
 use Illuminate\Http\Request;
+use Log;
 
 class PropertyController extends Controller
 {
@@ -67,13 +68,17 @@ class PropertyController extends Controller
     {
         /** @var Property $foundProperty */
         $foundProperty = Property::find($property->id);
-        if (!$foundProperty) {
+        if ( ! $foundProperty) {
             return response()->json(['message' => 'Property not found'], 404);
         }
 
-        $foundProperty->fill($request->input());
-        $foundProperty->update();
-        return response()->json(['property' => $foundProperty], 200);
+        if ($foundProperty->user_id == Auth::user()->getAuthIdentifier()) {
+            $foundProperty->fill($request->input());
+            $foundProperty->update();
+            return response()->json(['property' => $foundProperty], 200);
+        }
+
+        return response()->json(['message' => 'Unauthorized access. Cannot update.'], 401);
     }
 
     /**
@@ -91,7 +96,12 @@ class PropertyController extends Controller
         if (!$foundProperty) {
             return response()->json(['message' => 'Property not found'], 404);
         }
-        $foundProperty->delete();
-        return response()->json(['message' => "Property [$property->id] deleted."], 200);
+
+        if ($foundProperty->user_id == Auth::user()->getAuthIdentifier()) {
+            $foundProperty->delete();
+            return response()->json(['message' => "Property [$property->id] deleted."], 200);
+        }
+
+        return response()->json(['message' => 'Unauthorized access. Cannot delete.'], 401);
     }
 }
